Added separate agent class for all snmp pollers to use
